[Feature] 完成页面功能

+++ app/src/main/java/org/codebase/locationcheater/ui/dao/LocationDto.java
@@ -2,11 +2,11 @@ package org.codebase.locationcheater.ui.dao;
 
 public class LocationDto {
 
-    private float latitude;
+    private double latitude;
 
-    private float longitude;
+    private double longitude;
 
-    public LocationDto(float latitude, float longitude) {
+    public LocationDto(double latitude, double longitude) {
         this.latitude = latitude;
         this.longitude = longitude;
     }
@@ -14,19 +14,19 @@ public class LocationDto {
     public LocationDto() {
     }
 
-    public float getLatitude() {
+    public double getLatitude() {
         return latitude;
     }
 
-    public void setLatitude(float latitude) {
+    public void setLatitude(double latitude) {
         this.latitude = latitude;
     }
 
-    public float getLongitude() {
+    public double getLongitude() {
         return longitude;
     }
 
-    public void setLongitude(float longitude) {
+    public void setLongitude(double longitude) {
         this.longitude = longitude;
     }
 }
